requirements.txt
# Requirements for FastAPI application with SQL Server connectivity

# Core dependencies
python-dotenv==1.2.1

# Web framework
fastapi==0.121.3
uvicorn==0.38.0

# Database connectivity
sqlalchemy==2.0.44 # For database interactions
databases==0.9.0 # For async database support
pyodbc==5.3.0

mysql-connector-python==9.5.0 # MySQL database connector

# Data validation
pydantic==2.12.4 # For data validation and settings management
email-validator==2.3.0 # For validating email addresses with Pydantic